+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix `refreshTiles()` for vector tiles ([#5875](https://github.com/maplibre/maplibre-gl-js/pull/5875))
 - Revert changes to polygon intersection detection ([#5590](https://github.com/maplibre/maplibre-gl-js/pull/5590) caused issue [5864](https://github.com/maplibre/maplibre-gl-js/issues/5864))
 - Fix breaking clusters when non-integer value provided for `clusterMaxZoom` (warning is shown) ([#5929](https://github.com/maplibre/maplibre-gl-js/issues/5929)) + clarify API docs
+- Improve performance of mouse events with terrain by using WebGL2 async data readback ([2764](https://github.com/maplibre/maplibre-gl-js/issues/2764))
 
 
 ## 5.5.0